Create specialist pool, add specialist hire and promotion events

+++ Server/SubterfugeDatabaseProvider/Models/DbSpecialistConfiguration.cs
@@ -7,7 +7,7 @@ public class DbSpecialistConfiguration
     public string Id { get; set; } = Guid.NewGuid().ToString();
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public DateTime ExpiresAt { get; set; } = DateTime.MaxValue;
-    public long Priority { get; set; }
+    public int Priority { get; set; }
     public string SpecialistName { get; set; }
     public SimpleUser Creator { get; set; }
     public List<SpecialistEffectConfiguration> SpecialistEffects { get; set; }
